requirements.txt
anyio>=4.10.0
attrs>=25.3.0
beautifulsoup4>=4.13.4
blinker>=1.9.0
bs4>=0.0.2
CacheControl>=0.14.3
cachetools>=5.5.2
certifi>=2025.8.3
cffi>=1.17.1
charset-normalizer>=3.4.3
click>=8.2.1
colorama>=0.4.6
cryptography>=45.0.6
dnspython>=2.7.0
email_validator>=2.2.0
firebase_admin>=7.1.0
Flask>=3.1.2
google-api-core>=2.25.1
google-api-python-client>=2.179.0
google-auth>=2.40.3
google-auth-httplib2>=0.2.0
google-cloud-core>=2.4.3
google-cloud-firestore>=2.21.0
google-cloud-storage>=3.3.0
google-crc32c>=1.7.1
google-resumable-media>=2.7.2
googleapis-common-protos>=1.70.0
grpcio>=1.74.0
grpcio-status>=1.74.0
gunicorn>=23.0.0
h11>=0.16.0
h2>=4.2.0
hpack>=4.1.0
httpcore>=1.0.9
httplib2>=0.22.0
httpx>=0.28.1
hyperframe>=6.1.0
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.6
MarkupSafe>=3.0.2
msgpack>=1.1.1
outcome>=1.3.0.post0
packaging>=25.0
proto-plus>=1.26.1
protobuf>=6.32.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pycparser>=2.22
PyJWT>=2.10.1
pyparsing>=3.2.3
PySocks>=1.7.1
python-dotenv>=1.1.1
pytz>=2025.2
requests>=2.32.5
rsa>=4.9.1
selenium>=4.35.0
sniffio>=1.3.1
sortedcontainers>=2.4.0
soupsieve>=2.7
trio>=0.30.0
trio-websocket>=0.12.2
typing_extensions>=4.14.1
uritemplate>=4.2.0
urllib3>=2.5.0
websocket-client>=1.8.0
Werkzeug>=3.1.3
wsproto>=1.2.0
xmltodict>=0.14.2
